Added class DataManeger

+++ Assignment1/Models/Tank.cs
@@ -4,12 +4,9 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Volume { get; set; }
-        public decimal MaxVolume { get; set; }
+        public double Volume { get; set; }
+        public double MaxVolume { get; set; }
         public int UnitId { get; set; }
         public Unit Unit { get; set; }
-
     }
-
-
 }
